fix(compiler): Fix look up of entryComponents in AOT Summaries generated by references from router configuration.
Previously, when you attempted to bootstrap a component that had a router-outlet using ngsummaries, it would complain that the component was not provided by any module even if it was. This commit fixes a mistake (AFAICT) which caused the lookup of the component in the AOT summaries to fail.

I believe this change is safe. I've run the presubmit tests within google and there were no breakages caused by this change (although there were existing failures).